refactor: create database.ts and models.ts
