+++ index.ts
@@ -1,93 +1,5 @@
-import { Database } from "bun:sqlite";
-
-const db = new Database("chffr.sqlite", { create: true });
-
-db.run("PRAGMA journal_mode = WAL;");
-
-db.run(`
-  CREATE TABLE IF NOT EXISTS users (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    email TEXT UNIQUE NOT NULL,
-    provider TEXT NOT NULL,
-    provider_id TEXT NOT NULL,
-    name TEXT,
-    created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
-    updated_at DATETIME DEFAULT CURRENT_TIMESTAMP
-  );
-`);
-
-db.run(`
-  CREATE TABLE IF NOT EXISTS devices (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    dongle_id TEXT UNIQUE NOT NULL,
-    user_id INTEGER NOT NULL,
-    name TEXT,
-    created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
-    updated_at DATETIME DEFAULT CURRENT_TIMESTAMP,
-    FOREIGN KEY (user_id) REFERENCES users (id)
-  );
-`);
-
-class User {
-  id: number;
-  email: string;
-  provider: string;
-  provider_id: string;
-  name: string | null;
-  created_at: string;
-  updated_at: string;
-
-  static findById(id: number): User | null {
-    const query = db.query("SELECT * FROM users WHERE id = ?").as(User);
-    return query.get(id);
-  }
-
-  static findByProviderAndId(provider: string, provider_id: string): User | null {
-    const query = db.query("SELECT * FROM users WHERE provider = ? AND provider_id = ?").as(User);
-    return query.get(provider, provider_id);
-  }
-
-  static create(data: { email: string; provider: string; provider_id: string; name?: string }): User {
-    const query = db
-      .query(`
-      INSERT INTO users (email, provider, provider_id, name)
-      VALUES (?, ?, ?, ?)
-      RETURNING *
-    `)
-      .as(User);
-    return query.get(data.email, data.provider, data.provider_id, data.name || null);
-  }
-}
-
-class Device {
-  id: number;
-  dongle_id: string;
-  user_id: number;
-  name: string | null;
-  created_at: string;
-  updated_at: string;
-
-  static findByDongleId(dongle_id: string): Device | null {
-    const query = db.query("SELECT * FROM devices WHERE dongle_id = ?").as(Device);
-    return query.get(dongle_id);
-  }
-
-  static findByUserId(user_id: number): Device[] {
-    const query = db.query("SELECT * FROM devices WHERE user_id = ?").as(Device);
-    return query.all(user_id);
-  }
-
-  static create(data: { dongle_id: string; user_id: number; name?: string }): Device {
-    const query = db
-      .query(`
-      INSERT INTO devices (dongle_id, user_id, name)
-      VALUES (?, ?, ?)
-      RETURNING *
-    `)
-      .as(Device);
-    return query.get(data.dongle_id, data.user_id, data.name || null);
-  }
-}
+import "./db/database";
+import { Device, User } from "./db/models";
 
 const server = Bun.serve({
   port: 3000,
